requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
psycopg2==2.9.9
sqlparse==0.5.0
whitenoise==6.7.0
django-unfold==0.39.0
pillow==11.0.0